feat: collect leader data successfully
